Formatação utilizando a biblioteca Blue

+++ fib_x.pyx
@@ -1,4 +1,4 @@
-cpdef int fib(int n):
+cpdef unsigned long long int fib(int n):
     cdef int i
     cdef unsigned long long int a = 0
     cdef unsigned long long int b = 1
